requirements.txt
asgiref==3.5.0
asn1crypto==1.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
class-registry==2.1.2
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cssselect2==0.6.0
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-axes==5.32.0
django-filter==21.1
django-import-export==2.8.0
django-ipware==4.0.2
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-simple-captcha==0.5.17
django-simple-history==3.0.0
django-user-visit==0.5.1
django-useraudit==1.7.1
djangorestframework==3.13.1
gunicorn==20.1.0
imutils==0.5.4
matplotlib==3.5.1
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pyparsing==3.0.8
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
regex==2022.3.15
reportlab==3.6.9
requests==2.27.1
serializers==0.2.4
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tablib==3.2.1
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
ua-parser==0.10.0
Unipath==1.1
uritools==4.0.0
urllib3==1.26.9
user-agents==2.2.0
webencodings==0.5.1
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0